Pipeline tests time limitation reduced 100 times. However, there are still tests that require web queries that take up to 15 seconds to run.

+++ dff/core/pipeline/messenger_interface.py
@@ -65,7 +65,7 @@ class PollingMessengerInterface(MessengerInterface):
         self,
         pipeline_runner: PipelineRunnerFunction,
         loop: PollingProviderLoopFunction = lambda: True,
-        timeout: int = 0,
+        timeout: float = 0,
     ):
         """
         Method, running a request - response cycle in a loop.
